server/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
minio==7.2.20
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
django-mptt==0.18.0
django-elasticsearch-dsl==8.0
elasticsearch-dsl==8.18.0
elasticsearch==8.19.3